Add nice types of value
There is a "labelled" type: List [String "label"; value]
A "table" type: List [labelled1; labelled2; ...]

Two tables can be compared as ordered lists or unordered lists of
ordered elements, or uls of uls of ols, ...

A configuration on SGPT side is a table of options. An option is
labelled by a short name ("nplayers" is reserved). It is a table with
keys "name" (to display to end user), "type" which must be "int" for
now, and optionnally "min", "max", and/or "incr" for max value, min
value and increment. For instance, if min=1, max=5, incr=2, int 5 will
work but int 4 no.

A configuration on client side is a table of options which matches the
options of SGPT, except for "nplayers" which is computed by SGSJ. An
option is a labelled value. The type must match SGPT's corresponding
type and the value SGPT's min, max and increment.

The client will have to ask SGPT for its configuration to display a form
to the end-user.