Begining to incorporate QMake into ocamlbuild.
I have already guessed CMake configuration (where the incs and libs are,
where moc, uic, rcc and so on are). I will then put it into a
setup.data-like file, add tags, deps, flags and rules to enpower
ocamlbuild.

However, it might be difficult to enable dynamic dependancies for the
code generators... It may also be difficult to tell oasis to add a
target...